% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_meta.R
\name{get_category}
\alias{get_category}
\title{Extract appropriate \code{category} from \code{d_items} projected down by \code{linksto}}
\usage{
get_category(try_linksto, try_category)
}
\arguments{
\item{try_linksto}{a character string representing the events table for which a \code{category} determination is requested}

\item{try_category}{a character string representing the desired \code{category} code from the \code{d_items} table}
}
\value{
a character string representing the appropriate \code{category} code.
}
\description{
The function returns an appropriate \code{category} code for the specified \code{linksto} table indicating the events
table where the \code{d_items} for apply.  The routine attempts to ignore case and white space in determining the
appropriate \code{category} code.
}
\examples{
cevts_gen <- get_category("chartevents", "general")
cevts_gen
}
