% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_core.R
\name{m4_admissions}
\alias{m4_admissions}
\title{Provides base access to unique hospitalizations for patients in the database}
\usage{
m4_admissions(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{hadm_id})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

adm <- m4_admissions(con, cohort = 10137012)
dim(adm)

bigrquery::dbDisconnect(con)
}
