% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_icu.R
\name{m4_chartevents}
\alias{m4_chartevents}
\title{Access charted items occurring during the ICU stay}
\usage{
m4_chartevents(con, cohort = NULL, itemlist = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{itemlist}{an optional vector of item IDs defining the event types of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the chartevents table containing data representing all the charted data
available for a patient. During their ICU stay, the primary repository of a patient’s information is their
electronic chart. The electronic chart displays patients' routine vital signs and any additional information
relevant to their care: ventilator settings, laboratory values, code status, mental status, and so on.
As a result, the bulk of information about a patient’s stay is contained in chartevents.
}
\details{
Table attributes for chartevents table:

(\strong{PKEY} \code{subject_id}, \code{hadm_id}, \code{stay_id})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{stay_id}) -> icustays table

(\strong{FKEY} \code{itemid}) -> d_items table, chartevents sub-table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

evt <- m4_chartevents(con, cohort = 12384098)
dim(evt)

bigrquery::dbDisconnect(con)
}
