% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_dimensions.R
\name{m4_chartevents_items}
\alias{m4_chartevents_items}
\title{Provides base access to chartevent items reference}
\usage{
m4_chartevents_items(con)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{ITEMID})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

tab <- m4_chartevents_items(con)
tab

bigrquery::dbDisconnect(con)
}
