% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_dimensions.R
\name{m4_d_icd_diagnoses}
\alias{m4_d_icd_diagnoses}
\title{Access dimension table for diagnoses_icd; provides a description of ICD-9/ICD-10 billed diagnoses}
\usage{
m4_d_icd_diagnoses(con, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the d_icd_diagnoses table which defines concepts recorded in the diagnoses_icd
table in the hospital module.
}
\details{
(PKEY \code{icd_code}, \code{icd_version})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

tab <- m4_d_icd_diagnoses(con)
tab

bigrquery::dbDisconnect(con)
}
