% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_dimensions.R
\name{m4_d_labitems}
\alias{m4_d_labitems}
\title{Access dimension table for labevents; provides a description of all lab items}
\usage{
m4_d_labitems(con, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the d_labitems table which defines concepts recorded in the
labevents table in the hospital module.
}
\details{
Table attributes for d_labitems table:

(\strong{PKEY} \code{itemid}) reference data for labevents table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

tab <- m4_d_labitems(con)
tab

bigrquery::dbDisconnect(con)
}
