% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_icu.R
\name{m4_datetimeevents}
\alias{m4_datetimeevents}
\title{Provides base access to documented information which is in a date format (e.g. date of last dialysis).}
\usage{
m4_datetimeevents(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{subject_id}, \code{hadm_id}, \code{stay_id})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

evt <- m4_datetimeevents(con, cohort = 10137012)
evt

bigrquery::dbDisconnect(con)
}
