% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_diagnoses}
\alias{m4_diagnoses}
\title{Access billed ICD-9/ICD-10 diagnoses for hospitalizations}
\usage{
m4_diagnoses(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the diagnoses_icd table containing data which represent a record of all
diagnoses a patient was billed for during their hospital stay using the ICD-9 and ICD-10 ontologies. Diagnoses
are billed on hospital discharge, and are determined by trained persons who read signed clinical notes.
}
\details{
Table attributes for diagnoses_icd table:

(\strong{PKEY} \code{subject_id}, \code{hadm_id}, \code{seq_num})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{icd_code}, \code{icd_version}) -> d_icd_diagnoses table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

diag <- m4_diagnoses(con, cohort = 10137012)
dim(diag)

bigrquery::dbDisconnect(con)
}
