% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_diagnoses_icd}
\alias{m4_diagnoses_icd}
\title{Provides base access to billed ICD-9/ICD-10 diagnoses for hospitalizations}
\usage{
m4_diagnoses_icd(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{subject_id}, \code{hadm_id}, \code{seq_num})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

diag <- m4_diagnoses_icd(con, cohort = 10137012)
dim(diag)

bigrquery::dbDisconnect(con)
}
