% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_base.R
\name{m4_get_from_table}
\alias{m4_get_from_table}
\title{Simple generic SELECT * wrapper for extracting data from base tables}
\usage{
m4_get_from_table(con, table, where = NULL)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{table}{a character string representing a valid MIMIC-IV table name.}

\item{where}{an optional character string representing a SQL \code{WHERE} clause.}
}
\value{
a tibble with the results.
}
\description{
Function to perform a basic \code{SELECT *} from the noted table projected down by the optional
where clause.
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

m4_get_from_table(con, "d_items", where = "where itemid <= 220048")

bigrquery::dbDisconnect(con)
}
