% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_hcpcsevents}
\alias{m4_hcpcsevents}
\title{Access billed events occurring during the hospitalization; includes CPT codes}
\usage{
m4_hcpcsevents(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the hcpcsevents table containing data about all billed events occurring
during the hospitalization.
}
\details{
Table attributes for hcpcsevents table:

(\strong{PKEY} \code{subject_id}, \code{hadm_id}, \code{seq_num})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{hcpcs_cd}) -> d_hcpcs table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

evt <- m4_hcpcsevents(con, cohort = 10137012)
dim(evt)

bigrquery::dbDisconnect(con)
}
