% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_meta.R
\name{m4_meta_data}
\alias{m4_meta_data}
\title{Encapsulated meta data for MIMIC IV tables}
\usage{
m4_meta_data()
}
\value{
a tibble containing meta data concerning MIMIC IV tables.
}
\description{
The function returns a tibble with the table name, MIMIC module and a description (all provided from the
\href{https://mimic.mit.edu/docs/iv/}{MIMIC-IV website}).
}
\examples{
core <- m4_meta_data() \%>\%
    dplyr::filter(module == "mimic_core") \%>\%
    dplyr::transmute(statement = stringr::str_c("select * from physionet-data.", module, ".", table))
core
}
