% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_icu.R
\name{m4_outputevents}
\alias{m4_outputevents}
\title{Access information regarding patient outputs including urine, drainage, and so on}
\usage{
m4_outputevents(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the outputevents table containing data representing all information
regarding patient outputs, for example, urine, drainage, etc.
}
\details{
(PKEY \code{subject_id}, \code{hadm_id}, \code{stay_id})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

evt <- m4_outputevents(con, cohort = 12384098)
dim(evt)

bigrquery::dbDisconnect(con)
}
