% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_patterns.R
\name{m4_patient_admissions}
\alias{m4_patient_admissions}
\title{Access combined patient and admissions data.}
\usage{
m4_patient_admissions(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
The function calculates as part of the data frame
length of stay in the hospital in days (\code{los_hospital}) and age at admission in years (\code{admission_age}).
Patients who are older than 89 years old at any time in the database have had their date of birth
shifted to obscure their age and comply with HIPAA.  These ages appear in the data as >= 300.
As such, an additional field is added to categorize age into decades (\code{admission_decade}).
Patients older than 89 show up in the 90 and older age decade bucket.
}
\details{
This is patterned off the \code{icustay_detail} SQL in the mimic-code github repository referenced above.
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

patadm <- m4_patient_admissions(con, cohort = 16904137)
dim(patadm)

bigrquery::dbDisconnect(con)
}
