% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_patterns.R
\name{m4_patient_icustays}
\alias{m4_patient_icustays}
\title{Access combined patient, admission and icu stay data.}
\usage{
m4_patient_icustays(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This includes all of the information from the \code{mimic_get_patient_admissions} function plus addition
data on icu stays while in the hospital.
}
\details{
This is patterned off the \code{icustay_detail} SQL in the mimic-code github repository referenced above.
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

icudet <- m4_patient_icustays(con, cohort = 16904137)
dim(icudet)

bigrquery::dbDisconnect(con)
}
