% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_core.R
\name{m4_patients}
\alias{m4_patients}
\title{Access unique patients in the database}
\usage{
m4_patients(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the patients table containing data that is consistent for the
time span of a patient in the database - i.e., there is one unique entry per patient in the database.
}
\details{
Table attributes for patients table:

(\strong{PKEY} \code{subject_id})
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

pat <- m4_patients(con, cohort = 10137012)
dim(pat)

bigrquery::dbDisconnect(con)
}
