% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_poe}
\alias{m4_poe}
\title{Access orders made by providers relating to patient care}
\usage{
m4_poe(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the poe table containing data pertaining to provider order entry (POE)
which is the general interface through which care providers at the hospital enter orders. Most treatments
and procedures must be ordered via POE.
}
\details{
Table attributes for poe table:

(\strong{PKEY} \code{poe_id}, \code{poe_seq})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

poe <- m4_poe(con, cohort = 10137012)
dim(poe)

bigrquery::dbDisconnect(con)
}
