% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_poe_detail}
\alias{m4_poe_detail}
\title{Access supplementary information for orders made by providers in the hospital}
\usage{
m4_poe_detail(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the poe_detail table containing data representing further information
on POE orders. The table uses an Entity-Attribute-Value (EAV) model: the entity is poe_id, the attribute is
field_name, and the value is field_value. EAV tables allow for flexible description of entities when the
attributes are heterogenous.
}
\details{
Table attributes for poe_detail table:

(\strong{PKEY} \code{poe_id}, \code{poe_seq})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{poe_id}) -> poe table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

pod <- m4_poe_detail(con, cohort = 10137012)
dim(pod)

bigrquery::dbDisconnect(con)
}
