% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_prescriptions}
\alias{m4_prescriptions}
\title{Access prescribed medications}
\usage{
m4_prescriptions(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the prescriptions table containing data about prescribed medications.
Information includes the name of the drug, coded identifiers including the Generic Sequence Number (GSN) and
National Drug Code (NDC), the product strength, the formulary dose, and the route of administration.
}
\details{
(PKEY \code{subject_id}, \code{hadm_id}. \code{pharmacy_id})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

script <- m4_prescriptions(con, cohort = 10137012)
dim(script)

bigrquery::dbDisconnect(con)
}
