% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_dimensions.R
\name{m4_procedureevents_items}
\alias{m4_procedureevents_items}
\title{Access procedureevents items reference}
\usage{
m4_procedureevents_items(con, category = character())
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{category}{an character string defining the item category of interest}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the d_items table which defines concepts recorded in the procedureevents
table in the icu module.
}
\details{
(PKEY \code{ITEMID})
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

tab <- m4_procedureevents_items(con)
tab

bigrquery::dbDisconnect(con)
}
