% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_procedures}
\alias{m4_procedures}
\title{Access billed procedures for patients during their hospital stay}
\usage{
m4_procedures(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the procedures_icd table containing data which represents a record of
all procedures a patient was billed for during their hospital stay using the ICD-9 and ICD-10 ontologies.
}
\details{
(PKEY \code{subject_id}, \code{hadm_id}. \code{pharmacy_id})
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

icd <- m4_procedures(con, cohort = 10137012)
dim(icd)

bigrquery::dbDisconnect(con)
}
