% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_meta.R
\name{m4_service_decsriptions}
\alias{m4_service_decsriptions}
\title{Encapsulated descriptive data for services}
\usage{
m4_service_decsriptions()
}
\value{
a tibble containing descriptive data for services.
}
\description{
The function returns a tibble with descriptive data from services including service code, detailed description and
short description.
}
\examples{
neuro <- m4_service_decsriptions() \%>\%
    dplyr::filter(startsWith(short_description, "Neurologic")) \%>\%
    dplyr::select(service, short_description)
neuro
}
