% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMIC4db-package.R
\docType{package}
\name{MIMIC4db-package}
\alias{MIMIC4db}
\alias{MIMIC4db-package}
\title{MIMIC4db: Package for building, loading and referencing the Google BigQuery verion of the MIMIC-IV}
\description{
MIMIC4db provides a tightly bound set of routines to reference the Google BigQuery version of
the \href{https://physionet.org/content/mimiciii/1.0/}{MIMIC-IV v1.0} database using base access
routines from the \href{https://github.com/r-dbi/DBI}{DBI} R package with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.
}
\author{
\strong{Maintainer}: H. David Shea \email{hdshea@comcast.net} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
