% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_core.R
\name{m4_admissions}
\alias{m4_admissions}
\title{Access unique hospitalizations for patients in the database}
\usage{
m4_admissions(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the admissions table containing information regarding a patient’s
admission to the hospital. Since each unique hospital visit for a patient is assigned a unique \code{hadm_id},
the admissions table can be considered as a definition table for \code{hadm_id}. Information available includes
timing information for admission and discharge, demographic information, the source of the admission, and so on.
}
\details{
Table attributes for admissions table:

(\strong{PKEY} \code{hadm_id})

(\strong{FKEY} \code{subject_id}) -> patients table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

adm <- m4_admissions(con, cohort = 10137012)
dim(adm)

bigrquery::dbDisconnect(con)
}
