% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_dimensions.R
\name{m4_chartevents_items}
\alias{m4_chartevents_items}
\title{Access chartevents items reference}
\usage{
m4_chartevents_items(con)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the d_items table which defines concepts recorded in the chartevents table in
the icu module.
}
\details{
(PKEY \code{ITEMID})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

tab <- m4_chartevents_items(con)
tab

bigrquery::dbDisconnect(con)
}
