% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_dimensions.R
\name{m4_d_labitems}
\alias{m4_d_labitems}
\title{Provides base access to dimension table for labevents; provides a description of all lab items}
\usage{
m4_d_labitems(con, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{itemid})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

tab <- m4_d_labitems(con)
tab

bigrquery::dbDisconnect(con)
}
