% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_icu.R
\name{m4_datetimeevents}
\alias{m4_datetimeevents}
\title{Access documented information which is in a date format (e.g. date of last dialysis)}
\usage{
m4_datetimeevents(con, cohort = NULL, itemlist = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{itemlist}{an optional vector of item IDs defining the event types of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the datetimeevents table containing data which represents all
date measurements about a patient in the ICU.
}
\details{
For example, the date of last dialysis would be in the DATETIMEEVENTS table, but the systolic blood
pressure would not be in this table.

Table attributes for datetimeevents table:

(\strong{PKEY} \code{subject_id}, \code{hadm_id}, \code{stay_id})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{stay_id}) -> icustays table

(\strong{FKEY} \code{itemid}) -> d_items table, datetimeevents_items sub-table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

evt <- m4_datetimeevents(con, cohort = 15914798)
dim(evt)

bigrquery::dbDisconnect(con)
}
