% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_emar_detail}
\alias{m4_emar_detail}
\title{Access supplementary information for electronic administrations recorded in emar}
\usage{
m4_emar_detail(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{subject_id}, \code{emar_id}, \code{emar_seq}, \code{parent_field_ordinal})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

edet <- m4_emar_detail(con, cohort = 10137012)
dim(edet)

bigrquery::dbDisconnect(con)
}
