% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_emar_detail}
\alias{m4_emar_detail}
\title{Access supplementary information for electronic administrations recorded in emar}
\usage{
m4_emar_detail(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the emar_detail table containing data for each medicine administration
made in the EMAR table. Information includes the associated pharmacy order, the dose due, the dose given,
and many other parameters associated with the medical administration.
}
\details{
(PKEY \code{subject_id}, \code{emar_id}, \code{emar_seq}, \code{parent_field_ordinal})
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

edet <- m4_emar_detail(con, cohort = 10137012)
dim(edet)

bigrquery::dbDisconnect(con)
}
