% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_base.R
\name{m4_get_from_table}
\alias{m4_get_from_table}
\title{Simple generic SELECT * wrapper for extracting data from base tables}
\usage{
m4_get_from_table(con, table, where = NULL)
}
\arguments{
\item{con}{A DBIConnection object, as returned by dbConnect().}

\item{table}{a character string representing a valid MIMIC-IV table name.}

\item{where}{a character string representing a SQL \code{WHERE} clause.}
}
\value{
a tibble with the results.
}
\description{
Helper function to perform a basic \code{SELECT *} from the noted table projected down by the optional
where clause.
}
\examples{
con <- bigrquery::dbConnect(
   bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)
m4_get_from_table(con, "physionet-data.mimic_icu.d_items", where = "where itemid <= 223938")
bigrquery::dbDisconnect(con)
}
