% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_labevents}
\alias{m4_labevents}
\title{Access laboratory measurements sourced from patient derived specimens}
\usage{
m4_labevents(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the labevents table containing data pertaining to the results
of all laboratory measurements made for a single patient. These include hematology measurements,
blood gases, chemistry panels, and less common tests such as genetic assays.
}
\details{
Table attributes for labevents table:

(\strong{PKEY} \code{labevent_id})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{itemid}) -> d_labitems table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

evt <- m4_labevents(con, cohort = 10137012)
dim(evt)

bigrquery::dbDisconnect(con)
}
