% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_microbiologyevents}
\alias{m4_microbiologyevents}
\title{Access microbiology cultures}
\usage{
m4_microbiologyevents(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the microbiologyevents table containing data on all microbiology tests
which are a common procedure to check for infectious growth and to assess which antibiotic treatments are
most effective.
}
\details{
(PKEY \code{microevent_id})
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

evt <- m4_microbiologyevents(con, cohort = 10137012)
dim(evt)

bigrquery::dbDisconnect(con)
}
