% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_pharmacy}
\alias{m4_pharmacy}
\title{Access formulary, dosing, and other information for prescribed medications}
\usage{
m4_pharmacy(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the pharmacy table containing data of detailed information regarding
filled medications which were prescribed to the patient. Pharmacy information includes the dose of the drug,
the number of formulary doses, the frequency of dosing, the medication route, and the duration of the
prescription.
}
\details{
Table attributes for pharmacy table:

(\strong{PKEY} \code{pharmacy_id})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{poe_id}) -> poe table

(\strong{FKEY} \code{pharmacy_id}) -> prescriptions table

(\strong{FKEY} \code{pharmacy_id}) -> emar table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

evt <- m4_pharmacy(con, cohort = 10177976)
dim(evt)

bigrquery::dbDisconnect(con)
}
