% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_poe_detail}
\alias{m4_poe_detail}
\title{Provides base access to supplementary information for orders made by providers in the hospital}
\usage{
m4_poe_detail(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
(PKEY \code{subject_id}, \code{hadm_id}. \code{poe_seq})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

pod <- m4_poe_detail(con, cohort = 10137012)
dim(pod)

bigrquery::dbDisconnect(con)
}
