% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_prescriptions}
\alias{m4_prescriptions}
\title{Access prescribed medications}
\usage{
m4_prescriptions(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the prescriptions table containing data about prescribed medications.
Information includes the name of the drug, coded identifiers including the Generic Sequence Number (GSN) and
National Drug Code (NDC), the product strength, the formulary dose, and the route of administration.
}
\details{
Table attributes for prescriptions table:

(\strong{PKEY} \code{pharmacy_id})

(\strong{FKEY} \code{subject_id}) -> patients table

(\strong{FKEY} \code{hadm_id}) -> admissions table

(\strong{FKEY} \code{pharmacy_id}) -> pharmacy table

(\strong{FKEY} \code{pharmacy_id}) -> emar table
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

script <- m4_prescriptions(con, cohort = 10177976)
dim(script)

bigrquery::dbDisconnect(con)
}
