% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_icu.R
\name{m4_procedureevents}
\alias{m4_procedureevents}
\title{Access procedures documented during the ICU stay}
\usage{
m4_procedureevents(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object with an appropriate
\href{https://github.com/r-dbi/bigrquery}{bigrquery} DBIConnection.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This functions provides base access to the procedureevents table from the hospital icu.
They include data on all procedures documented during the ICU stay (e.g. ventilation), though not
necessarily conducted within the ICU (e.g. x-ray imaging).
}
\details{
(PKEY \code{subject_id}, \code{hadm_id}, \code{stay_id})
}
\examples{
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

evt <- m4_procedureevents(con, cohort = 12384098)
dim(evt)

bigrquery::dbDisconnect(con)
}
