% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_base.R
\name{m4_select_data}
\alias{m4_select_data}
\title{Basic SELECT statement wrapper returning results in a tibble}
\usage{
m4_select_data(con, select_statement)
}
\arguments{
\item{con}{A DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{select_statement}{a character string representing a SQL \code{SELECT} statement.}
}
\value{
a tibble with the results.
}
\description{
Wrapper around DBI functions to send a \code{SELECT} statement, fetch the results
and return them in a tibble.
}
\examples{
x <- 5
}
