% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_base.R
\name{m4_select_data}
\alias{m4_select_data}
\title{Basic SELECT statement wrapper returning results in a tibble}
\usage{
m4_select_data(con, select_statement)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{select_statement}{a character string representing a SQL \code{SELECT} statement.}
}
\value{
a tibble with the results.
}
\description{
Wrapper around DBI functions to send a \code{SELECT} statement, fetch the results
and return them in a tibble.
}
\examples{
# To run examples, you must have the BIGQUERY_TEST_PROJECT environment
# variable set to name of project which has billing set up and to which
# you have write access.
con <- bigrquery::dbConnect(
    bigrquery::bigquery(),
    project = bigrquery::bq_test_project(),
    quiet = TRUE
)

m4_select_data(con, "select * from physionet-data.mimic_icu.d_items order by itemid limit 5")

bigrquery::dbDisconnect(con)
}
