% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m4_hosp.R
\name{m4_services}
\alias{m4_services}
\title{Access the hospital service(s) which cared for the patient during their hospitalization}
\usage{
m4_services(con, cohort = NULL, ...)
}
\arguments{
\item{con}{A \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBIConnection object, as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
with an appropriate \code{\link[bigrquery:bigquery]{bigrquery::bigquery()}} DBI driver specified in the call.}

\item{cohort}{an optional vector of patient IDs defining the cohort of interest}

\item{...}{additional optional passed along parameters.}
}
\value{
a tibble with the results.
}
\description{
This function provides base access to the services table containing data that describes the service under which
a patient was admitted.
}
\details{
(PKEY \code{subject_id}, \code{hadm_id}. \code{pharmacy_id})
}
\examples{
con <- bigrquery::dbConnect(
  bigrquery::bigquery(),
  project = bigrquery::bq_test_project(),
  quiet = TRUE
)

svs <- m4_services(con, cohort = 10137012)
dim(svs)

bigrquery::dbDisconnect(con)
}
